# BigDecimal: roundSignificand perf improvement (when using scale)