@ionspin 2024-07-14
:tada: :tada: :tada: Thank you very much for all the pull requests! Sorry for the delay in merging them, I'll try to publish a stable release soon!